Add more ceph release abstraction and use utils scripts in undercloud install
This abstracts the ceph release entirely out of the playbooks, and
puts it in group_vars where it belongs.

This commit also adds some use of the utils helper scripts to let
ansible determine if it should run or skip certain tasks, which
hopefully increases idempotence.

I.E. Don't try to build overcloud glance images if they already exist.