Add image lookup for undercloud image install and fix baremetal lookup
Adds a task to make sure the images directory is created.

Adds a call to the check_service.py script to check whether images
have already been build and installed.

Fixes an error with the baremetal provision task where it was set
to run if the baremetal server could is greater than zero instead
of equal to 0.